Struct aws_sdk_lambda::client::fluent_builders::CreateFunction

pub struct CreateFunction { /* fields omitted */ }

Implementations

impl CreateFunction

pub async fn send(
self
) -> Result<CreateFunctionOutput, SdkError<CreateFunctionError>>

pub fn function_name(self, inp: impl Into<String>) -> Self

The name of the Lambda function.

Name formats

Function name - my-function.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

pub fn set_function_name(self, inp: Option<String>) -> Self

pub fn runtime(self, inp: Runtime) -> Self

The identifier of the function's runtime.

pub fn set_runtime(self, inp: Option<Runtime>) -> Self

pub fn role(self, inp: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the function's execution role.

pub fn set_role(self, inp: Option<String>) -> Self

pub fn handler(self, inp: impl Into<String>) -> Self

The name of the method within your code that Lambda calls to execute your function. The format includes the file name. It can also include namespaces and other qualifiers, depending on the runtime. For more information, see Programming Model.

pub fn set_handler(self, inp: Option<String>) -> Self

pub fn code(self, inp: FunctionCode) -> Self

The code for the function.

pub fn set_code(self, inp: Option<FunctionCode>) -> Self

pub fn description(self, inp: impl Into<String>) -> Self

A description of the function.

pub fn set_description(self, inp: Option<String>) -> Self

pub fn timeout(self, inp: i32) -> Self

The amount of time that Lambda allows a function to run before stopping it. The default is 3 seconds. The maximum allowed value is 900 seconds.

pub fn set_timeout(self, inp: Option<i32>) -> Self

pub fn memory_size(self, inp: i32) -> Self

The amount of memory available to the function at runtime. Increasing the function's memory also increases its CPU allocation. The default value is 128 MB. The value can be any multiple of 1 MB.

pub fn set_memory_size(self, inp: Option<i32>) -> Self

pub fn publish(self, inp: bool) -> Self

Set to true to publish the first version of the function during creation.

pub fn set_publish(self, inp: bool) -> Self

pub fn vpc_config(self, inp: VpcConfig) -> Self

For network connectivity to AWS resources in a VPC, specify a list of security groups and subnets in the VPC. When you connect a function to a VPC, it can only access resources and the internet through that VPC. For more information, see VPC Settings.

pub fn set_vpc_config(self, inp: Option<VpcConfig>) -> Self

pub fn package_type(self, inp: PackageType) -> Self

The type of deployment package. Set to Image for container image and set Zip for ZIP archive.

pub fn set_package_type(self, inp: Option<PackageType>) -> Self

pub fn dead_letter_config(self, inp: DeadLetterConfig) -> Self

A dead letter queue configuration that specifies the queue or topic where Lambda sends asynchronous events when they fail processing. For more information, see Dead Letter Queues.

pub fn set_dead_letter_config(self, inp: Option<DeadLetterConfig>) -> Self

pub fn environment(self, inp: Environment) -> Self

Environment variables that are accessible from function code during execution.

pub fn set_environment(self, inp: Option<Environment>) -> Self

pub fn kms_key_arn(self, inp: impl Into<String>) -> Self

The ARN of the AWS Key Management Service (AWS KMS) key that's used to encrypt your function's environment variables. If it's not provided, AWS Lambda uses a default service key.

pub fn set_kms_key_arn(self, inp: Option<String>) -> Self

pub fn tracing_config(self, inp: TracingConfig) -> Self

Set Mode to Active to sample and trace a subset of incoming requests with AWS X-Ray.

pub fn set_tracing_config(self, inp: Option<TracingConfig>) -> Self

pub fn tags(self, k: impl Into<String>, v: impl Into<String>) -> Self

A list of tags to apply to the function.

pub fn set_tags(self, inp: Option<HashMap<String, String>>) -> Self

pub fn layers(self, inp: impl Into<String>) -> Self

A list of function layers to add to the function's execution environment. Specify each layer by its ARN, including the version.

pub fn set_layers(self, inp: Option<Vec<String>>) -> Self

pub fn file_system_configs(self, inp: impl Into<FileSystemConfig>) -> Self

Connection settings for an Amazon EFS file system.

pub fn set_file_system_configs(self, inp: Option<Vec<FileSystemConfig>>) -> Self

pub fn image_config(self, inp: ImageConfig) -> Self

Container image configuration values that override the values in the container image Dockerfile.

pub fn set_image_config(self, inp: Option<ImageConfig>) -> Self

pub fn code_signing_config_arn(self, inp: impl Into<String>) -> Self

To enable code signing for this function, specify the ARN of a code-signing configuration. A code-signing configuration includes a set of signing profiles, which define the trusted publishers for this function.

pub fn set_code_signing_config_arn(self, inp: Option<String>) -> Self