Struct aws_sdk_lambda::client::fluent_builders::UpdateFunctionConfiguration

pub struct UpdateFunctionConfiguration<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> UpdateFunctionConfiguration<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<UpdateFunctionConfigurationOutput, SdkError<UpdateFunctionConfigurationError>> where
R::Policy: SmithyRetryPolicy<UpdateFunctionConfigurationInputOperationOutputAlias, UpdateFunctionConfigurationOutput, UpdateFunctionConfigurationError, UpdateFunctionConfigurationInputOperationRetryAlias>,

pub fn function_name(self, inp: impl Into<String>) -> Self

The name of the Lambda function.

Name formats

Function name - my-function.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

pub fn set_function_name(self, input: Option<String>) -> Self

pub fn role(self, inp: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the function's execution role.

pub fn set_role(self, input: Option<String>) -> Self

pub fn handler(self, inp: impl Into<String>) -> Self

The name of the method within your code that Lambda calls to execute your function. The format includes the file name. It can also include namespaces and other qualifiers, depending on the runtime. For more information, see Programming Model.

pub fn set_handler(self, input: Option<String>) -> Self

pub fn description(self, inp: impl Into<String>) -> Self

A description of the function.

pub fn set_description(self, input: Option<String>) -> Self

pub fn timeout(self, inp: i32) -> Self

The amount of time that Lambda allows a function to run before stopping it. The default is 3 seconds. The maximum allowed value is 900 seconds. For additional information, see Lambda execution environment.

pub fn set_timeout(self, input: Option<i32>) -> Self

pub fn memory_size(self, inp: i32) -> Self

The amount of memory available to the function at runtime. Increasing the function memory also increases its CPU allocation. The default value is 128 MB. The value can be any multiple of 1 MB.

pub fn set_memory_size(self, input: Option<i32>) -> Self

pub fn vpc_config(self, inp: VpcConfig) -> Self

For network connectivity to Amazon Web Services resources in a VPC, specify a list of security groups and subnets in the VPC. When you connect a function to a VPC, it can only access resources and the internet through that VPC. For more information, see VPC Settings.

pub fn set_vpc_config(self, input: Option<VpcConfig>) -> Self

pub fn environment(self, inp: Environment) -> Self

Environment variables that are accessible from function code during execution.

pub fn set_environment(self, input: Option<Environment>) -> Self

pub fn runtime(self, inp: Runtime) -> Self

The identifier of the function's runtime.

pub fn set_runtime(self, input: Option<Runtime>) -> Self

pub fn dead_letter_config(self, inp: DeadLetterConfig) -> Self

A dead letter queue configuration that specifies the queue or topic where Lambda sends asynchronous events when they fail processing. For more information, see Dead Letter Queues.

pub fn set_dead_letter_config(self, input: Option<DeadLetterConfig>) -> Self

pub fn kms_key_arn(self, inp: impl Into<String>) -> Self

The ARN of the Amazon Web Services Key Management Service (KMS) key that's used to encrypt your function's environment variables. If it's not provided, Lambda uses a default service key.

pub fn set_kms_key_arn(self, input: Option<String>) -> Self

pub fn tracing_config(self, inp: TracingConfig) -> Self

Set Mode to Active to sample and trace a subset of incoming requests with X-Ray.

pub fn set_tracing_config(self, input: Option<TracingConfig>) -> Self

pub fn revision_id(self, inp: impl Into<String>) -> Self

Only update the function if the revision ID matches the ID that's specified. Use this option to avoid modifying a function that has changed since you last read it.

pub fn set_revision_id(self, input: Option<String>) -> Self

pub fn layers(self, inp: impl Into<String>) -> Self

Appends an item to Layers.

To override the contents of this collection use set_layers.

A list of function layers to add to the function's execution environment. Specify each layer by its ARN, including the version.

pub fn set_layers(self, input: Option<Vec<String>>) -> Self

pub fn file_system_configs(self, inp: impl Into<FileSystemConfig>) -> Self

Appends an item to FileSystemConfigs.

To override the contents of this collection use set_file_system_configs.

Connection settings for an Amazon EFS file system.

pub fn set_file_system_configs(
self,
input: Option<Vec<FileSystemConfig>>
) -> Self

pub fn image_config(self, inp: ImageConfig) -> Self

Container image configuration values that override the values in the container image Dockerfile.

pub fn set_image_config(self, input: Option<ImageConfig>) -> Self