Module aws_sdk_lambda::error

Modules

code_signing_config_not_found_exception

See CodeSigningConfigNotFoundException

code_storage_exceeded_exception

See CodeStorageExceededException

code_verification_failed_exception

See CodeVerificationFailedException

ec2_access_denied_exception

See Ec2AccessDeniedException

ec2_throttled_exception

See Ec2ThrottledException

ec2_unexpected_exception

See Ec2UnexpectedException

efs_mount_connectivity_exception

See EfsMountConnectivityException

efs_mount_failure_exception

See EfsMountFailureException

efs_mount_timeout_exception

See EfsMountTimeoutException

efsio_exception

See EfsioException

eni_limit_reached_exception

See EniLimitReachedException

invalid_code_signature_exception

See InvalidCodeSignatureException

invalid_parameter_value_exception

See InvalidParameterValueException

invalid_request_content_exception

See InvalidRequestContentException

invalid_runtime_exception

See InvalidRuntimeException

invalid_security_group_id_exception

See InvalidSecurityGroupIdException

invalid_subnet_id_exception

See InvalidSubnetIdException

invalid_zip_file_exception

See InvalidZipFileException

kms_access_denied_exception

See KmsAccessDeniedException

kms_disabled_exception

See KmsDisabledException

kms_invalid_state_exception

See KmsInvalidStateException

kms_not_found_exception

See KmsNotFoundException

policy_length_exceeded_exception

See PolicyLengthExceededException

precondition_failed_exception

See PreconditionFailedException

provisioned_concurrency_config_not_found_exception

See ProvisionedConcurrencyConfigNotFoundException

request_too_large_exception

See RequestTooLargeException

resource_conflict_exception

See ResourceConflictException

resource_in_use_exception

See ResourceInUseException

resource_not_found_exception

See ResourceNotFoundException

resource_not_ready_exception

See ResourceNotReadyException

service_exception

See ServiceException

subnet_ip_address_limit_reached_exception

See SubnetIpAddressLimitReachedException

too_many_requests_exception

See TooManyRequestsException

unsupported_media_type_exception

See UnsupportedMediaTypeException

Structs

AddLayerVersionPermissionError

AddPermissionError

CodeSigningConfigNotFoundException

The specified code signing configuration does not exist.

CodeStorageExceededException

You have exceeded your maximum total code size per account. Learn more

CodeVerificationFailedException

The code signature failed one or more of the validation checks for signature mismatch or expiry, and the code signing policy

CreateAliasError

CreateCodeSigningConfigError

CreateEventSourceMappingError

CreateFunctionError

DeleteAliasError

DeleteCodeSigningConfigError

DeleteEventSourceMappingError

DeleteFunctionCodeSigningConfigError

DeleteFunctionConcurrencyError

DeleteFunctionError

DeleteFunctionEventInvokeConfigError

DeleteLayerVersionError

DeleteProvisionedConcurrencyConfigError

Ec2AccessDeniedException

Need additional permissions to configure VPC settings.

Ec2ThrottledException

AWS Lambda was throttled by Amazon EC2 during Lambda function initialization using the execution role provided

Ec2UnexpectedException

AWS Lambda received an unexpected EC2 client exception while setting up for the Lambda function.

EfsMountConnectivityException

The function couldn't make a network connection to the configured file system.

EfsMountFailureException

The function couldn't mount the configured file system due to a permission or configuration issue.

EfsMountTimeoutException

The function was able to make a network connection to the configured file system, but the mount operation

EfsioException

An error occured when reading from or writing to a connected file system.

EniLimitReachedException

AWS Lambda was not able to create an elastic network interface in the VPC, specified as part of Lambda

GetAccountSettingsError

GetAliasError

GetCodeSigningConfigError

GetEventSourceMappingError

GetFunctionCodeSigningConfigError

GetFunctionConcurrencyError

GetFunctionConfigurationError

GetFunctionError

GetFunctionEventInvokeConfigError

GetLayerVersionByArnError

GetLayerVersionError

GetLayerVersionPolicyError

GetPolicyError

GetProvisionedConcurrencyConfigError

InvalidCodeSignatureException

The code signature failed the integrity check. Lambda always blocks deployment if the integrity check

InvalidParameterValueException

One of the parameters in the request is invalid.

InvalidRequestContentException

The request body could not be parsed as JSON.

InvalidRuntimeException

The runtime or runtime version specified is not supported.

InvalidSecurityGroupIdException

The Security Group ID provided in the Lambda function VPC configuration is invalid.

InvalidSubnetIdException

The Subnet ID provided in the Lambda function VPC configuration is invalid.

InvalidZipFileException

AWS Lambda could not unzip the deployment package.

InvokeAsyncError

InvokeError

KmsAccessDeniedException

Lambda was unable to decrypt the environment variables because KMS access was denied. Check the Lambda

KmsDisabledException

Lambda was unable to decrypt the environment variables because the KMS key used is disabled. Check the Lambda

KmsInvalidStateException

Lambda was unable to decrypt the environment variables because the KMS key used is in an invalid state for

KmsNotFoundException

Lambda was unable to decrypt the environment variables because the KMS key was not found. Check the function's

ListAliasesError

ListCodeSigningConfigsError

ListEventSourceMappingsError

ListFunctionEventInvokeConfigsError

ListFunctionsByCodeSigningConfigError

ListFunctionsError

ListLayerVersionsError

ListLayersError

ListProvisionedConcurrencyConfigsError

ListTagsError

ListVersionsByFunctionError

PolicyLengthExceededException

The permissions policy for the resource is too large. Learn more

PreconditionFailedException

The RevisionId provided does not match the latest RevisionId for the Lambda function or alias. Call the

ProvisionedConcurrencyConfigNotFoundException

The specified configuration does not exist.

PublishLayerVersionError

PublishVersionError

PutFunctionCodeSigningConfigError

PutFunctionConcurrencyError

PutFunctionEventInvokeConfigError

PutProvisionedConcurrencyConfigError

RemoveLayerVersionPermissionError

RemovePermissionError

RequestTooLargeException

The request payload exceeded the Invoke request body JSON input limit. For more information, see

ResourceConflictException

The resource already exists, or another operation is in progress.

ResourceInUseException

The operation conflicts with the resource's availability. For example, you attempted to update an EventSource

ResourceNotFoundException

The resource specified in the request does not exist.

ResourceNotReadyException

The function is inactive and its VPC connection is no longer available. Wait for the VPC connection to

ServiceException

The AWS Lambda service encountered an internal error.

SubnetIpAddressLimitReachedException

AWS Lambda was not able to set up VPC access for the Lambda function because one or more configured subnets

TagResourceError

TooManyRequestsException

The request throughput limit was exceeded.

UnsupportedMediaTypeException

The content type of the Invoke request body is not JSON.

UntagResourceError

UpdateAliasError

UpdateCodeSigningConfigError

UpdateEventSourceMappingError

UpdateFunctionCodeError

UpdateFunctionConfigurationError

UpdateFunctionEventInvokeConfigError

Enums

AddLayerVersionPermissionErrorKind

AddPermissionErrorKind

CreateAliasErrorKind

CreateCodeSigningConfigErrorKind

CreateEventSourceMappingErrorKind

CreateFunctionErrorKind

DeleteAliasErrorKind

DeleteCodeSigningConfigErrorKind

DeleteEventSourceMappingErrorKind

DeleteFunctionCodeSigningConfigErrorKind

DeleteFunctionConcurrencyErrorKind

DeleteFunctionErrorKind

DeleteFunctionEventInvokeConfigErrorKind

DeleteLayerVersionErrorKind

DeleteProvisionedConcurrencyConfigErrorKind

GetAccountSettingsErrorKind

GetAliasErrorKind

GetCodeSigningConfigErrorKind

GetEventSourceMappingErrorKind

GetFunctionCodeSigningConfigErrorKind

GetFunctionConcurrencyErrorKind

GetFunctionConfigurationErrorKind

GetFunctionErrorKind

GetFunctionEventInvokeConfigErrorKind

GetLayerVersionByArnErrorKind

GetLayerVersionErrorKind

GetLayerVersionPolicyErrorKind

GetPolicyErrorKind

GetProvisionedConcurrencyConfigErrorKind

InvokeAsyncErrorKind

InvokeErrorKind

ListAliasesErrorKind

ListCodeSigningConfigsErrorKind

ListEventSourceMappingsErrorKind

ListFunctionEventInvokeConfigsErrorKind

ListFunctionsByCodeSigningConfigErrorKind

ListFunctionsErrorKind

ListLayerVersionsErrorKind

ListLayersErrorKind

ListProvisionedConcurrencyConfigsErrorKind

ListTagsErrorKind

ListVersionsByFunctionErrorKind

PublishLayerVersionErrorKind

PublishVersionErrorKind

PutFunctionCodeSigningConfigErrorKind

PutFunctionConcurrencyErrorKind

PutFunctionEventInvokeConfigErrorKind

PutProvisionedConcurrencyConfigErrorKind

RemoveLayerVersionPermissionErrorKind

RemovePermissionErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateAliasErrorKind

UpdateCodeSigningConfigErrorKind

UpdateEventSourceMappingErrorKind

UpdateFunctionCodeErrorKind

UpdateFunctionConfigurationErrorKind

UpdateFunctionEventInvokeConfigErrorKind