Module aws_sdk_lambda::operation

Structs

AddLayerVersionPermission

Adds permissions to the resource-based policy of a version of an AWS Lambda

AddPermission

Grants an AWS service or another account permission to use a function. You can apply the policy at the

CreateAlias

Creates an alias for a

CreateCodeSigningConfig

Creates a code signing configuration. A code signing configuration defines a list of

CreateEventSourceMapping

Creates a mapping between an event source and an AWS Lambda function. Lambda reads items from the event source

CreateFunction

Creates a Lambda function. To create a function, you need a deployment package and an execution role. The

DeleteAlias

Deletes a Lambda function alias.

DeleteCodeSigningConfig

Deletes the code signing configuration. You can delete the code signing configuration only if no function is

DeleteEventSourceMapping

Deletes an event source

DeleteFunction

Deletes a Lambda function. To delete a specific function version, use the Qualifier parameter.

DeleteFunctionCodeSigningConfig

Removes the code signing configuration from the function.

DeleteFunctionConcurrency

Removes a concurrent execution limit from a function.

DeleteFunctionEventInvokeConfig

Deletes the configuration for asynchronous invocation for a function, version, or alias.

DeleteLayerVersion

Deletes a version of an AWS Lambda

DeleteProvisionedConcurrencyConfig

Deletes the provisioned concurrency configuration for a function.

GetAccountSettings

Retrieves details about your account's limits and usage in an AWS Region.

GetAlias

Returns details about a Lambda function alias.

GetCodeSigningConfig

Returns information about the specified code signing configuration.

GetEventSourceMapping

Returns details about an event source mapping. You can get the identifier of a mapping from the output of ListEventSourceMappings.

GetFunction

Returns information about the function or function version, with a link to download the deployment package

GetFunctionCodeSigningConfig

Returns the code signing configuration for the specified function.

GetFunctionConcurrency

Returns details about the reserved concurrency configuration for a function. To set a concurrency limit for a

GetFunctionConfiguration

Returns the version-specific settings of a Lambda function or version. The output includes only options that

GetFunctionEventInvokeConfig

Retrieves the configuration for asynchronous invocation for a function, version, or alias.

GetLayerVersion

Returns information about a version of an AWS Lambda

GetLayerVersionByArn

Returns information about a version of an AWS Lambda

GetLayerVersionPolicy

Returns the permission policy for a version of an AWS Lambda

GetPolicy

Returns the resource-based IAM policy for a function, version, or alias.

GetProvisionedConcurrencyConfig

Retrieves the provisioned concurrency configuration for a function's alias or version.

Invoke

Invokes a Lambda function. You can invoke a function synchronously (and wait for the response), or

InvokeAsync

ListAliases

Returns a list of aliases

ListCodeSigningConfigs

Returns a list of code

ListEventSourceMappings

Lists event source mappings. Specify an EventSourceArn to only show event source mappings for a

ListFunctionEventInvokeConfigs

Retrieves a list of configurations for asynchronous invocation for a function.

ListFunctions

Returns a list of Lambda functions, with the version-specific configuration of each. Lambda returns up to 50

ListFunctionsByCodeSigningConfig

List the functions that use the specified code signing configuration. You can use this method prior to deleting a

ListLayerVersions

Lists the versions of an AWS Lambda

ListLayers

Lists AWS Lambda

ListProvisionedConcurrencyConfigs

Retrieves a list of provisioned concurrency configurations for a function.

ListTags

Returns a function's tags. You can

ListVersionsByFunction

Returns a list of versions,

PublishLayerVersion

Creates an AWS Lambda

PublishVersion

Creates a version from the

PutFunctionCodeSigningConfig

Update the code signing configuration for the function. Changes to the code signing configuration take effect the

PutFunctionConcurrency

Sets the maximum number of simultaneous executions for a function, and reserves capacity for that concurrency

PutFunctionEventInvokeConfig

Configures options for asynchronous

PutProvisionedConcurrencyConfig

Adds a provisioned concurrency configuration to a function's alias or version.

RemoveLayerVersionPermission

Removes a statement from the permissions policy for a version of an AWS Lambda

RemovePermission

Revokes function-use permission from an AWS service or another account. You can get the ID of the statement

TagResource

Adds tags to a function.

UntagResource

Removes tags from a function.

UpdateAlias

Updates the configuration of a Lambda function alias.

UpdateCodeSigningConfig

Update the code signing configuration. Changes to the code signing configuration take effect the next time a

UpdateEventSourceMapping

Updates an event source mapping. You can change the function that AWS Lambda invokes, or pause invocation and resume later from the same location.

UpdateFunctionCode

Updates a Lambda function's code. If code signing is enabled for the function, the code package must be signed

UpdateFunctionConfiguration

Modify the version-specific settings of a Lambda function.

UpdateFunctionEventInvokeConfig

Updates the configuration for asynchronous invocation for a function, version, or alias.