Struct aws_sdk_lambda::output::create_function_output::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for CreateFunctionOutput

Implementations

impl Builder

pub fn function_name(self, input: impl Into<String>) -> Self

The name of the function.

pub fn set_function_name(self, input: Option<String>) -> Self

pub fn function_arn(self, input: impl Into<String>) -> Self

The function's Amazon Resource Name (ARN).

pub fn set_function_arn(self, input: Option<String>) -> Self

pub fn runtime(self, input: Runtime) -> Self

The runtime environment for the Lambda function.

pub fn set_runtime(self, input: Option<Runtime>) -> Self

pub fn role(self, input: impl Into<String>) -> Self

The function's execution role.

pub fn set_role(self, input: Option<String>) -> Self

pub fn handler(self, input: impl Into<String>) -> Self

The function that Lambda calls to begin executing your function.

pub fn set_handler(self, input: Option<String>) -> Self

pub fn code_size(self, input: i64) -> Self

The size of the function's deployment package, in bytes.

pub fn set_code_size(self, input: Option<i64>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

The function's description.

pub fn set_description(self, input: Option<String>) -> Self

pub fn timeout(self, input: i32) -> Self

The amount of time in seconds that Lambda allows a function to run before stopping it.

pub fn set_timeout(self, input: Option<i32>) -> Self

pub fn memory_size(self, input: i32) -> Self

The amount of memory available to the function at runtime.

pub fn set_memory_size(self, input: Option<i32>) -> Self

pub fn last_modified(self, input: impl Into<String>) -> Self

The date and time that the function was last updated, in ISO-8601 format (YYYY-MM-DDThh:mm:ss.sTZD).

pub fn set_last_modified(self, input: Option<String>) -> Self

pub fn code_sha256(self, input: impl Into<String>) -> Self

The SHA256 hash of the function's deployment package.

pub fn set_code_sha256(self, input: Option<String>) -> Self

pub fn version(self, input: impl Into<String>) -> Self

The version of the Lambda function.

pub fn set_version(self, input: Option<String>) -> Self

pub fn vpc_config(self, input: VpcConfigResponse) -> Self

The function's networking configuration.

pub fn set_vpc_config(self, input: Option<VpcConfigResponse>) -> Self

pub fn dead_letter_config(self, input: DeadLetterConfig) -> Self

The function's dead letter queue.

pub fn set_dead_letter_config(self, input: Option<DeadLetterConfig>) -> Self

pub fn environment(self, input: EnvironmentResponse) -> Self

The function's environment variables.

pub fn set_environment(self, input: Option<EnvironmentResponse>) -> Self

pub fn kms_key_arn(self, input: impl Into<String>) -> Self

The KMS key that's used to encrypt the function's environment variables. This key is only returned if you've configured a customer managed CMK.

pub fn set_kms_key_arn(self, input: Option<String>) -> Self

pub fn tracing_config(self, input: TracingConfigResponse) -> Self

The function's AWS X-Ray tracing configuration.

pub fn set_tracing_config(self, input: Option<TracingConfigResponse>) -> Self

pub fn master_arn(self, input: impl Into<String>) -> Self

For Lambda@Edge functions, the ARN of the master function.

pub fn set_master_arn(self, input: Option<String>) -> Self

pub fn revision_id(self, input: impl Into<String>) -> Self

The latest updated revision of the function or alias.

pub fn set_revision_id(self, input: Option<String>) -> Self

pub fn layers(self, input: impl Into<Layer>) -> Self

pub fn set_layers(self, input: Option<Vec<Layer>>) -> Self

pub fn state(self, input: State) -> Self

The current state of the function. When the state is Inactive, you can reactivate the function by invoking it.

pub fn set_state(self, input: Option<State>) -> Self

pub fn state_reason(self, input: impl Into<String>) -> Self

The reason for the function's current state.

pub fn set_state_reason(self, input: Option<String>) -> Self

pub fn state_reason_code(self, input: StateReasonCode) -> Self

The reason code for the function's current state. When the code is Creating, you can't invoke or modify the function.

pub fn set_state_reason_code(self, input: Option<StateReasonCode>) -> Self

pub fn last_update_status(self, input: LastUpdateStatus) -> Self

The status of the last update that was performed on the function. This is first set to Successful after function creation completes.

pub fn set_last_update_status(self, input: Option<LastUpdateStatus>) -> Self

pub fn last_update_status_reason(self, input: impl Into<String>) -> Self

The reason for the last update that was performed on the function.

pub fn set_last_update_status_reason(self, input: Option<String>) -> Self

pub fn last_update_status_reason_code(
self,
input: LastUpdateStatusReasonCode
) -> Self

The reason code for the last update that was performed on the function.

pub fn set_last_update_status_reason_code(
self,
input: Option<LastUpdateStatusReasonCode>
) -> Self

pub fn file_system_configs(self, input: impl Into<FileSystemConfig>) -> Self

pub fn set_file_system_configs(
self,
input: Option<Vec<FileSystemConfig>>
) -> Self

pub fn package_type(self, input: PackageType) -> Self

The type of deployment package. Set to Image for container image and set Zip for .zip file archive.

pub fn set_package_type(self, input: Option<PackageType>) -> Self

pub fn image_config_response(self, input: ImageConfigResponse) -> Self

The function's image configuration values.

pub fn set_image_config_response(
self,
input: Option<ImageConfigResponse>
) -> Self

pub fn signing_profile_version_arn(self, input: impl Into<String>) -> Self

The ARN of the signing profile version.

pub fn set_signing_profile_version_arn(self, input: Option<String>) -> Self

pub fn signing_job_arn(self, input: impl Into<String>) -> Self

The ARN of the signing job.

pub fn set_signing_job_arn(self, input: Option<String>) -> Self

pub fn build(self) -> CreateFunctionOutput

Consumes the builder and constructs a CreateFunctionOutput