Struct aws_sdk_lambda::output::delete_event_source_mapping_output::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for DeleteEventSourceMappingOutput

Implementations

impl Builder

pub fn uuid(self, input: impl Into<String>) -> Self

The identifier of the event source mapping.

pub fn set_uuid(self, input: Option<String>) -> Self

pub fn starting_position(self, input: EventSourcePosition) -> Self

The position in a stream from which to start reading. Required for Amazon Kinesis, Amazon DynamoDB, and Amazon MSK Streams sources. AT_TIMESTAMP is only supported for Amazon Kinesis streams.

pub fn set_starting_position(self, input: Option<EventSourcePosition>) -> Self

pub fn starting_position_timestamp(self, input: Instant) -> Self

With StartingPosition set to AT_TIMESTAMP, the time from which to start reading.

pub fn set_starting_position_timestamp(self, input: Option<Instant>) -> Self

pub fn batch_size(self, input: i32) -> Self

The maximum number of items to retrieve in a single batch.

pub fn set_batch_size(self, input: Option<i32>) -> Self

pub fn maximum_batching_window_in_seconds(self, input: i32) -> Self

(Streams and SQS standard queues) The maximum amount of time to gather records before invoking the function, in seconds. The default value is zero.

pub fn set_maximum_batching_window_in_seconds(self, input: Option<i32>) -> Self

pub fn parallelization_factor(self, input: i32) -> Self

(Streams) The number of batches to process from each shard concurrently. The default value is 1.

pub fn set_parallelization_factor(self, input: Option<i32>) -> Self

pub fn event_source_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the event source.

pub fn set_event_source_arn(self, input: Option<String>) -> Self

pub fn function_arn(self, input: impl Into<String>) -> Self

The ARN of the Lambda function.

pub fn set_function_arn(self, input: Option<String>) -> Self

pub fn last_modified(self, input: Instant) -> Self

The date that the event source mapping was last updated, or its state changed.

pub fn set_last_modified(self, input: Option<Instant>) -> Self

pub fn last_processing_result(self, input: impl Into<String>) -> Self

The result of the last AWS Lambda invocation of your Lambda function.

pub fn set_last_processing_result(self, input: Option<String>) -> Self

pub fn state(self, input: impl Into<String>) -> Self

The state of the event source mapping. It can be one of the following: Creating, Enabling, Enabled, Disabling, Disabled, Updating, or Deleting.

pub fn set_state(self, input: Option<String>) -> Self

pub fn state_transition_reason(self, input: impl Into<String>) -> Self

Indicates whether the last change to the event source mapping was made by a user, or by the Lambda service.

pub fn set_state_transition_reason(self, input: Option<String>) -> Self

pub fn destination_config(self, input: DestinationConfig) -> Self

(Streams) An Amazon SQS queue or Amazon SNS topic destination for discarded records.

pub fn set_destination_config(self, input: Option<DestinationConfig>) -> Self

pub fn topics(self, input: impl Into<String>) -> Self

pub fn set_topics(self, input: Option<Vec<String>>) -> Self

pub fn queues(self, input: impl Into<String>) -> Self

pub fn set_queues(self, input: Option<Vec<String>>) -> Self

pub fn source_access_configurations(
self,
input: impl Into<SourceAccessConfiguration>
) -> Self

pub fn set_source_access_configurations(
self,
input: Option<Vec<SourceAccessConfiguration>>
) -> Self

pub fn self_managed_event_source(self, input: SelfManagedEventSource) -> Self

The Self-Managed Apache Kafka cluster for your event source.

pub fn set_self_managed_event_source(
self,
input: Option<SelfManagedEventSource>
) -> Self

pub fn maximum_record_age_in_seconds(self, input: i32) -> Self

(Streams) Discard records older than the specified age. The default value is infinite (-1). When set to infinite (-1), failed records are retried until the record expires.

pub fn set_maximum_record_age_in_seconds(self, input: Option<i32>) -> Self

pub fn bisect_batch_on_function_error(self, input: bool) -> Self

(Streams) If the function returns an error, split the batch in two and retry. The default value is false.

pub fn set_bisect_batch_on_function_error(self, input: Option<bool>) -> Self

pub fn maximum_retry_attempts(self, input: i32) -> Self

(Streams) Discard records after the specified number of retries. The default value is infinite (-1). When set to infinite (-1), failed records are retried until the record expires.

pub fn set_maximum_retry_attempts(self, input: Option<i32>) -> Self

pub fn tumbling_window_in_seconds(self, input: i32) -> Self

(Streams) The duration in seconds of a processing window. The range is between 1 second up to 900 seconds.

pub fn set_tumbling_window_in_seconds(self, input: Option<i32>) -> Self

pub fn function_response_types(
self,
input: impl Into<FunctionResponseType>
) -> Self

pub fn set_function_response_types(
self,
input: Option<Vec<FunctionResponseType>>
) -> Self

pub fn build(self) -> DeleteEventSourceMappingOutput

Consumes the builder and constructs a DeleteEventSourceMappingOutput