Struct aws_sdk_lambda::output::DeleteEventSourceMappingOutput

#[non_exhaustive]
pub struct DeleteEventSourceMappingOutput {

    pub uuid: Option<String>,
    pub starting_position: Option<EventSourcePosition>,
    pub starting_position_timestamp: Option<Instant>,
    pub batch_size: Option<i32>,
    pub maximum_batching_window_in_seconds: Option<i32>,
    pub parallelization_factor: Option<i32>,
    pub event_source_arn: Option<String>,
    pub function_arn: Option<String>,
    pub last_modified: Option<Instant>,
    pub last_processing_result: Option<String>,
    pub state: Option<String>,
    pub state_transition_reason: Option<String>,
    pub destination_config: Option<DestinationConfig>,
    pub topics: Option<Vec<String>>,
    pub queues: Option<Vec<String>>,
    pub source_access_configurations: Option<Vec<SourceAccessConfiguration>>,
    pub self_managed_event_source: Option<SelfManagedEventSource>,
    pub maximum_record_age_in_seconds: Option<i32>,
    pub bisect_batch_on_function_error: Option<bool>,
    pub maximum_retry_attempts: Option<i32>,
    pub tumbling_window_in_seconds: Option<i32>,
    pub function_response_types: Option<Vec<FunctionResponseType>>,

}

A mapping between an AWS resource and an AWS Lambda function. See CreateEventSourceMapping for details.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

uuid: Option<String>

The identifier of the event source mapping.

starting_position: Option<EventSourcePosition>

The position in a stream from which to start reading. Required for Amazon Kinesis, Amazon DynamoDB, and Amazon MSK Streams sources. AT_TIMESTAMP is only supported for Amazon Kinesis streams.

starting_position_timestamp: Option<Instant>

With StartingPosition set to AT_TIMESTAMP, the time from which to start reading.

batch_size: Option<i32>

The maximum number of items to retrieve in a single batch.

maximum_batching_window_in_seconds: Option<i32>

(Streams and SQS standard queues) The maximum amount of time to gather records before invoking the function, in seconds. The default value is zero.

parallelization_factor: Option<i32>

(Streams) The number of batches to process from each shard concurrently. The default value is 1.

event_source_arn: Option<String>

The Amazon Resource Name (ARN) of the event source.

function_arn: Option<String>

The ARN of the Lambda function.

last_modified: Option<Instant>

The date that the event source mapping was last updated, or its state changed.

last_processing_result: Option<String>

The result of the last AWS Lambda invocation of your Lambda function.

state: Option<String>

The state of the event source mapping. It can be one of the following: Creating, Enabling, Enabled, Disabling, Disabled, Updating, or Deleting.

state_transition_reason: Option<String>

Indicates whether the last change to the event source mapping was made by a user, or by the Lambda service.

destination_config: Option<DestinationConfig>

(Streams) An Amazon SQS queue or Amazon SNS topic destination for discarded records.

topics: Option<Vec<String>>

The name of the Kafka topic.

queues: Option<Vec<String>>

(MQ) The name of the Amazon MQ broker destination queue to consume.

source_access_configurations: Option<Vec<SourceAccessConfiguration>>

An array of the authentication protocol, or the VPC components to secure your event source.

self_managed_event_source: Option<SelfManagedEventSource>

The Self-Managed Apache Kafka cluster for your event source.

maximum_record_age_in_seconds: Option<i32>

(Streams) Discard records older than the specified age. The default value is infinite (-1). When set to infinite (-1), failed records are retried until the record expires.

bisect_batch_on_function_error: Option<bool>

(Streams) If the function returns an error, split the batch in two and retry. The default value is false.

maximum_retry_attempts: Option<i32>

(Streams) Discard records after the specified number of retries. The default value is infinite (-1). When set to infinite (-1), failed records are retried until the record expires.

tumbling_window_in_seconds: Option<i32>

(Streams) The duration in seconds of a processing window. The range is between 1 second up to 900 seconds.

function_response_types: Option<Vec<FunctionResponseType>>

(Streams) A list of current response type enums applied to the event source mapping.

Implementations

impl DeleteEventSourceMappingOutput

pub fn builder() -> Builder

Creates a new builder-style object to manufacture DeleteEventSourceMappingOutput