Struct aws_sdk_lambda::output::UpdateFunctionCodeOutput

#[non_exhaustive]
pub struct UpdateFunctionCodeOutput {

    pub function_name: Option<String>,
    pub function_arn: Option<String>,
    pub runtime: Option<Runtime>,
    pub role: Option<String>,
    pub handler: Option<String>,
    pub code_size: i64,
    pub description: Option<String>,
    pub timeout: Option<i32>,
    pub memory_size: Option<i32>,
    pub last_modified: Option<String>,
    pub code_sha256: Option<String>,
    pub version: Option<String>,
    pub vpc_config: Option<VpcConfigResponse>,
    pub dead_letter_config: Option<DeadLetterConfig>,
    pub environment: Option<EnvironmentResponse>,
    pub kms_key_arn: Option<String>,
    pub tracing_config: Option<TracingConfigResponse>,
    pub master_arn: Option<String>,
    pub revision_id: Option<String>,
    pub layers: Option<Vec<Layer>>,
    pub state: Option<State>,
    pub state_reason: Option<String>,
    pub state_reason_code: Option<StateReasonCode>,
    pub last_update_status: Option<LastUpdateStatus>,
    pub last_update_status_reason: Option<String>,
    pub last_update_status_reason_code: Option<LastUpdateStatusReasonCode>,
    pub file_system_configs: Option<Vec<FileSystemConfig>>,
    pub package_type: Option<PackageType>,
    pub image_config_response: Option<ImageConfigResponse>,
    pub signing_profile_version_arn: Option<String>,
    pub signing_job_arn: Option<String>,

}

Details about a function's configuration.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

function_name: Option<String>

The name of the function.

function_arn: Option<String>

The function's Amazon Resource Name (ARN).

runtime: Option<Runtime>

The runtime environment for the Lambda function.

role: Option<String>

The function's execution role.

handler: Option<String>

The function that Lambda calls to begin executing your function.

code_size: i64

The size of the function's deployment package, in bytes.

description: Option<String>

The function's description.

timeout: Option<i32>

The amount of time in seconds that Lambda allows a function to run before stopping it.

memory_size: Option<i32>

The amount of memory available to the function at runtime.

last_modified: Option<String>

The date and time that the function was last updated, in ISO-8601 format (YYYY-MM-DDThh:mm:ss.sTZD).

code_sha256: Option<String>

The SHA256 hash of the function's deployment package.

version: Option<String>

The version of the Lambda function.

vpc_config: Option<VpcConfigResponse>

The function's networking configuration.

dead_letter_config: Option<DeadLetterConfig>

The function's dead letter queue.

environment: Option<EnvironmentResponse>

The function's environment variables.

kms_key_arn: Option<String>

The KMS key that's used to encrypt the function's environment variables. This key is only returned if you've configured a customer managed CMK.

tracing_config: Option<TracingConfigResponse>

The function's AWS X-Ray tracing configuration.

master_arn: Option<String>

For Lambda@Edge functions, the ARN of the master function.

revision_id: Option<String>

The latest updated revision of the function or alias.

layers: Option<Vec<Layer>>

The function's layers.

state: Option<State>

The current state of the function. When the state is Inactive, you can reactivate the function by invoking it.

state_reason: Option<String>

The reason for the function's current state.

state_reason_code: Option<StateReasonCode>

The reason code for the function's current state. When the code is Creating, you can't invoke or modify the function.

last_update_status: Option<LastUpdateStatus>

The status of the last update that was performed on the function. This is first set to Successful after function creation completes.

last_update_status_reason: Option<String>

The reason for the last update that was performed on the function.

last_update_status_reason_code: Option<LastUpdateStatusReasonCode>

The reason code for the last update that was performed on the function.

file_system_configs: Option<Vec<FileSystemConfig>>

Connection settings for an Amazon EFS file system.

package_type: Option<PackageType>

The type of deployment package. Set to Image for container image and set Zip for .zip file archive.

image_config_response: Option<ImageConfigResponse>

The function's image configuration values.

signing_profile_version_arn: Option<String>

The ARN of the signing profile version.

signing_job_arn: Option<String>

The ARN of the signing job.

Implementations

impl UpdateFunctionCodeOutput

pub fn builder() -> Builder

Creates a new builder-style object to manufacture UpdateFunctionCodeOutput