Module aws_sdk_lambda::error

Modules

code_signing_config_not_found_error
See CodeSigningConfigNotFoundError

code_storage_exceeded_error
See CodeStorageExceededError

code_verification_failed_error
See CodeVerificationFailedError

ec2_access_denied_error
See EC2AccessDeniedError

ec2_throttled_error
See EC2ThrottledError

ec2_unexpected_error
See EC2UnexpectedError

efs_mount_connectivity_error
See EFSMountConnectivityError

efs_mount_failure_error
See EFSMountFailureError

efs_mount_timeout_error
See EFSMountTimeoutError

efsio_error
See EFSIOError

eni_limit_reached_error
See ENILimitReachedError

invalid_code_signature_error
See InvalidCodeSignatureError

invalid_parameter_value_error
See InvalidParameterValueError

invalid_request_content_error
See InvalidRequestContentError

invalid_runtime_error
See InvalidRuntimeError

invalid_security_group_id_error
See InvalidSecurityGroupIDError

invalid_subnet_id_error
See InvalidSubnetIDError

invalid_zip_file_error
See InvalidZipFileError

kms_access_denied_error
See KMSAccessDeniedError

kms_disabled_error
See KMSDisabledError

kms_invalid_state_error
See KMSInvalidStateError

kms_not_found_error
See KMSNotFoundError

policy_length_exceeded_error
See PolicyLengthExceededError

precondition_failed_error
See PreconditionFailedError

provisioned_concurrency_config_not_found_error
See ProvisionedConcurrencyConfigNotFoundError

request_too_large_error
See RequestTooLargeError

resource_conflict_error
See ResourceConflictError

resource_in_use_error
See ResourceInUseError

resource_not_found_error
See ResourceNotFoundError

resource_not_ready_error
See ResourceNotReadyError

service_error
See ServiceError

subnet_ip_address_limit_reached_error
See SubnetIPAddressLimitReachedError

too_many_requests_error
See TooManyRequestsError

unsupported_media_type_error
See UnsupportedMediaTypeError

Structs

AddLayerVersionPermissionError

AddPermissionError

CodeSigningConfigNotFoundError
The specified code signing configuration does not exist.

CodeStorageExceededError
You have exceeded your maximum total code size per account. Learn more

CodeVerificationFailedError
The code signature failed one or more of the validation checks for signature mismatch or expiry, and the code signing policy

CreateAliasError

CreateCodeSigningConfigError

CreateEventSourceMappingError

CreateFunctionError

DeleteAliasError

DeleteCodeSigningConfigError

DeleteEventSourceMappingError

DeleteFunctionCodeSigningConfigError

DeleteFunctionConcurrencyError

DeleteFunctionError

DeleteFunctionEventInvokeConfigError

DeleteLayerVersionError

DeleteProvisionedConcurrencyConfigError

EC2AccessDeniedError
Need additional permissions to configure VPC settings.

EC2ThrottledError
AWS Lambda was throttled by Amazon EC2 during Lambda function initialization using the execution role provided

EC2UnexpectedError
AWS Lambda received an unexpected EC2 client exception while setting up for the Lambda function.

EFSIOError
An error occured when reading from or writing to a connected file system.

EFSMountConnectivityError
The function couldn't make a network connection to the configured file system.

EFSMountFailureError
The function couldn't mount the configured file system due to a permission or configuration issue.

EFSMountTimeoutError
The function was able to make a network connection to the configured file system, but the mount operation

ENILimitReachedError
AWS Lambda was not able to create an elastic network interface in the VPC, specified as part of Lambda

GetAccountSettingsError

GetAliasError

GetCodeSigningConfigError

GetEventSourceMappingError

GetFunctionCodeSigningConfigError

GetFunctionConcurrencyError

GetFunctionConfigurationError

GetFunctionError

GetFunctionEventInvokeConfigError

GetLayerVersionByArnError

GetLayerVersionError

GetLayerVersionPolicyError

GetPolicyError

GetProvisionedConcurrencyConfigError

InvalidCodeSignatureError
The code signature failed the integrity check. Lambda always blocks deployment if the integrity check

InvalidParameterValueError
One of the parameters in the request is invalid.

InvalidRequestContentError
The request body could not be parsed as JSON.

InvalidRuntimeError
The runtime or runtime version specified is not supported.

InvalidSecurityGroupIDError
The Security Group ID provided in the Lambda function VPC configuration is invalid.

InvalidSubnetIDError
The Subnet ID provided in the Lambda function VPC configuration is invalid.

InvalidZipFileError
AWS Lambda could not unzip the deployment package.

InvokeAsyncError

InvokeError

KMSAccessDeniedError
Lambda was unable to decrypt the environment variables because KMS access was denied. Check the Lambda

KMSDisabledError
Lambda was unable to decrypt the environment variables because the KMS key used is disabled. Check the Lambda

KMSInvalidStateError
Lambda was unable to decrypt the environment variables because the KMS key used is in an invalid state for

KMSNotFoundError
Lambda was unable to decrypt the environment variables because the KMS key was not found. Check the function's

ListAliasesError

ListCodeSigningConfigsError

ListEventSourceMappingsError

ListFunctionEventInvokeConfigsError

ListFunctionsByCodeSigningConfigError

ListFunctionsError

ListLayerVersionsError

ListLayersError

ListProvisionedConcurrencyConfigsError

ListTagsError

ListVersionsByFunctionError

PolicyLengthExceededError
The permissions policy for the resource is too large. Learn more

PreconditionFailedError
The RevisionId provided does not match the latest RevisionId for the Lambda function or alias. Call the

ProvisionedConcurrencyConfigNotFoundError
The specified configuration does not exist.

PublishLayerVersionError

PublishVersionError

PutFunctionCodeSigningConfigError

PutFunctionConcurrencyError

PutFunctionEventInvokeConfigError

PutProvisionedConcurrencyConfigError

RemoveLayerVersionPermissionError

RemovePermissionError

RequestTooLargeError
The request payload exceeded the Invoke request body JSON input limit. For more information, see

ResourceConflictError
The resource already exists, or another operation is in progress.

ResourceInUseError
The operation conflicts with the resource's availability. For example, you attempted to update an EventSource

ResourceNotFoundError
The resource specified in the request does not exist.

ResourceNotReadyError
The function is inactive and its VPC connection is no longer available. Wait for the VPC connection to

ServiceError
The AWS Lambda service encountered an internal error.

SubnetIPAddressLimitReachedError
AWS Lambda was not able to set up VPC access for the Lambda function because one or more configured subnets

TagResourceError

TooManyRequestsError
The request throughput limit was exceeded.

UnsupportedMediaTypeError
The content type of the Invoke request body is not JSON.

UntagResourceError

UpdateAliasError

UpdateCodeSigningConfigError

UpdateEventSourceMappingError

UpdateFunctionCodeError

UpdateFunctionConfigurationError

UpdateFunctionEventInvokeConfigError

Enums

AddLayerVersionPermissionErrorKind

AddPermissionErrorKind

CreateAliasErrorKind

CreateCodeSigningConfigErrorKind

CreateEventSourceMappingErrorKind

CreateFunctionErrorKind

DeleteAliasErrorKind

DeleteCodeSigningConfigErrorKind

DeleteEventSourceMappingErrorKind

DeleteFunctionCodeSigningConfigErrorKind

DeleteFunctionConcurrencyErrorKind

DeleteFunctionErrorKind

DeleteFunctionEventInvokeConfigErrorKind

DeleteLayerVersionErrorKind

DeleteProvisionedConcurrencyConfigErrorKind

GetAccountSettingsErrorKind

GetAliasErrorKind

GetCodeSigningConfigErrorKind

GetEventSourceMappingErrorKind

GetFunctionCodeSigningConfigErrorKind

GetFunctionConcurrencyErrorKind

GetFunctionConfigurationErrorKind

GetFunctionErrorKind

GetFunctionEventInvokeConfigErrorKind

GetLayerVersionByArnErrorKind

GetLayerVersionErrorKind

GetLayerVersionPolicyErrorKind

GetPolicyErrorKind

GetProvisionedConcurrencyConfigErrorKind

InvokeAsyncErrorKind

InvokeErrorKind

ListAliasesErrorKind

ListCodeSigningConfigsErrorKind

ListEventSourceMappingsErrorKind

ListFunctionEventInvokeConfigsErrorKind

ListFunctionsByCodeSigningConfigErrorKind

ListFunctionsErrorKind

ListLayerVersionsErrorKind

ListLayersErrorKind

ListProvisionedConcurrencyConfigsErrorKind

ListTagsErrorKind

ListVersionsByFunctionErrorKind

PublishLayerVersionErrorKind

PublishVersionErrorKind

PutFunctionCodeSigningConfigErrorKind

PutFunctionConcurrencyErrorKind

PutFunctionEventInvokeConfigErrorKind

PutProvisionedConcurrencyConfigErrorKind

RemoveLayerVersionPermissionErrorKind

RemovePermissionErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateAliasErrorKind

UpdateCodeSigningConfigErrorKind

UpdateEventSourceMappingErrorKind

UpdateFunctionCodeErrorKind

UpdateFunctionConfigurationErrorKind

UpdateFunctionEventInvokeConfigErrorKind