Module aws_sdk_lambda::model

Modules

account_limit
See AccountLimit

account_usage
See AccountUsage

alias_configuration
See AliasConfiguration

alias_routing_configuration
See AliasRoutingConfiguration

allowed_publishers
See AllowedPublishers

code_signing_config
See CodeSigningConfig

code_signing_policies
See CodeSigningPolicies

concurrency
See Concurrency

dead_letter_config
See DeadLetterConfig

destination_config
See DestinationConfig

environment
See Environment

environment_error
See EnvironmentError

environment_response
See EnvironmentResponse

event_source_mapping_configuration
See EventSourceMappingConfiguration

file_system_config
See FileSystemConfig

function_code
See FunctionCode

function_code_location
See FunctionCodeLocation

function_configuration
See FunctionConfiguration

function_event_invoke_config
See FunctionEventInvokeConfig

image_config
See ImageConfig

image_config_error
See ImageConfigError

image_config_response
See ImageConfigResponse

layer
See Layer

layer_version_content_input
See LayerVersionContentInput

layer_version_content_output
See LayerVersionContentOutput

layer_versions_list_item
See LayerVersionsListItem

layers_list_item
See LayersListItem

on_failure
See OnFailure

on_success
See OnSuccess

provisioned_concurrency_config_list_item
See ProvisionedConcurrencyConfigListItem

self_managed_event_source
See SelfManagedEventSource

source_access_configuration
See SourceAccessConfiguration

tracing_config
See TracingConfig

tracing_config_response
See TracingConfigResponse

vpc_config
See VpcConfig

vpc_config_response
See VpcConfigResponse

Structs

AccountLimit
Limits that are related to concurrency and storage. All file and storage sizes are in bytes.

AccountUsage
The number of functions and amount of storage in use.

AliasConfiguration
Provides configuration information about a Lambda function alias.

AliasRoutingConfiguration
The traffic-shifting configuration of a Lambda function alias.

AllowedPublishers
List of signing profiles that can sign a code package.

CodeSigningConfig
Details about a Code signing configuration.

CodeSigningPolicies
Code signing configuration policies specifies the validation failure action for signature mismatch or

Concurrency

DeadLetterConfig
The dead-letter queue for

DestinationConfig
A configuration object that specifies the destination of an event after Lambda processes it.

Environment
A function's environment variable settings.

EnvironmentError
Error messages for environment variables that couldn't be applied.

EnvironmentResponse
The results of an operation to update or read environment variables. If the operation is successful, the

EventSourceMappingConfiguration
A mapping between an AWS resource and an AWS Lambda function. See CreateEventSourceMapping for details.

FileSystemConfig
Details about the connection between a Lambda function and an Amazon EFS file system.

FunctionCode
The code for the Lambda function. You can specify either an object in Amazon S3, upload a .zip file archive deployment

FunctionCodeLocation
Details about a function's deployment package.

FunctionConfiguration
Details about a function's configuration.

FunctionEventInvokeConfig

ImageConfig
Configuration values that override the container image Dockerfile settings. See

ImageConfigError
Error response to GetFunctionConfiguration.

ImageConfigResponse
Response to GetFunctionConfiguration request.

Layer
An AWS Lambda

LayerVersionContentInput
A ZIP archive that contains the contents of an AWS Lambda

LayerVersionContentOutput
Details about a version of an AWS Lambda

LayerVersionsListItem
Details about a version of an AWS Lambda

LayersListItem
Details about an AWS Lambda

OnFailure
A destination for events that failed processing.

OnSuccess
A destination for events that were processed successfully.

ProvisionedConcurrencyConfigListItem
Details about the provisioned concurrency configuration for a function alias or version.

SelfManagedEventSource
The Self-Managed Apache Kafka cluster for your event source.

SourceAccessConfiguration
You can specify the authentication protocol, or the VPC components to secure access to your event source.

TracingConfig
The function's AWS X-Ray tracing configuration. To sample and record incoming requests, set Mode

TracingConfigResponse
The function's AWS X-Ray tracing configuration.

VpcConfig
The VPC security groups and subnets that are attached to a Lambda function. For more information, see VPC Settings.

VpcConfigResponse
The VPC security groups and subnets that are attached to a Lambda function.

Enums

CodeSigningPolicy

EndPointType

EventSourcePosition

FunctionResponseType

FunctionVersion

InvocationType

LastUpdateStatus

LastUpdateStatusReasonCode

LogType

PackageType

ProvisionedConcurrencyStatusEnum

Runtime

SourceAccessType

State

StateReasonCode

ThrottleReason

TracingMode